requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
typing-extensions==3.7.4.3
Werkzeug==1.0.1
